organized post-performance pdf report into separate program
